dual_farming: remove lock xToken to active farming
