--- tests/dual-farming/keys/prog-TeSTKchdpa2FKNV6gYNAENpququb3aT2r1pD41tZw36.json
@@ -1,1 +1,0 @@
-[99,88,159,24,211,86,41,8,220,47,170,72,113,8,235,248,163,245,155,116,136,37,0,234,29,195,51,32,107,84,14,97,6,211,110,235,255,101,133,147,157,75,55,231,230,2,30,58,113,250,121,172,228,125,95,21,9,28,194,76,18,158,209,1]
