--- tests/dual-farming/keys/xstep-tEsTL8G8drugWztoCKrPpEAXV21qEajfHg4q45KYs6s.json
@@ -1,1 +1,0 @@
-[189,65,246,88,57,128,70,1,53,105,104,67,55,128,210,188,81,117,180,44,137,15,53,94,186,60,127,245,98,220,241,179,13,32,77,211,54,104,98,71,147,164,12,225,147,190,146,162,246,165,174,16,157,170,173,120,22,111,46,157,85,216,114,148]
